DRY: Move helper functions to helpers module
