+++ src/ct_clip/scripts/zero_shot.py
@@ -14,6 +14,7 @@ from torch.utils.data import DataLoader
 from transformers import BertTokenizer
 
 from ct_clip import CTCLIP, get_optimizer
+from ct_clip.helpers import noop, cycle
 from data_inference import CTReportDatasetinfer
 from eval import evaluate_internal
 
@@ -42,20 +43,6 @@ def tensor_to_nifti(tensor, path, affine=np.eye(4)):
     numpy_data = tensor.detach().numpy().astype(np.float32)
     nifti_img = nib.Nifti1Image(numpy_data, affine)
     nib.save(nifti_img, path)
-
-
-def exists(val):
-    return val is not None
-
-
-def noop(*args, **kwargs):
-    pass
-
-
-def cycle(dl):
-    while True:
-        for data in dl:
-            yield data
 
 
 def yes_or_no(question):
